Fix macOS and Windows build errors
Replace GNU ##__VA_ARGS__ extension in LOG macro with standard C++20
__VA_OPT__ to fix Clang -Wpedantic error on macOS. Add static_cast
for int-to-WORD narrowing in actions.cpp to fix MSVC /W4 /WX error.

Made-with: Cursor

+++ src/actions.cpp
@@ -96,7 +96,7 @@ bool actions::register_all(reaper_plugin_info_t* rec) {
     g_cmd_id = rec->Register("command_id", (void*)"ReaperDemucs_SeparateSources");
     if (!g_cmd_id) return false;
 
-    g_accel.accel.cmd = g_cmd_id;
+    g_accel.accel.cmd = static_cast<WORD>(g_cmd_id);
     g_accel.accel.fVirt = 0;
     g_accel.accel.key = 0;
     g_accel.desc = "Separate stems";
